Add support for fractional difficulty in mining.set_difficulty

+++ components/stratum/include/stratum_api.h
@@ -65,7 +65,7 @@ typedef struct
     // mining.notify
     mining_notify *mining_notification;
     // mining.set_difficulty
-    uint32_t new_difficulty;
+    double new_difficulty;
     // mining.set_version_mask
     uint32_t version_mask;
     // result
